requirements.txt
tensorflow>=2.0.0
pandas>=1.0.0
numpy>=1.18.0
matplotlib>=3.0.0
scikit-learn>=0.22.0
psycopg2-binary>=2.8.0
python-dotenv>=0.19.0

